crypto-producer/requirements.txt
kafka-python
faker
ccxt